Remove switch container volume mount dev/log
This mount was used to gather logs from containers and log them in the
execution machine, it was an old practice from first topology days and
is not required anymore.

The removal is mainly because there are some vtysh commands that extract
data from local log files, so the logs need to be on the container instead
of the execution machine.